+++ pkg/common/observability_bundle.go
@@ -5,13 +5,14 @@ import (
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/blang/semver"
 	appv1 "github.com/giantswarm/apiextensions-application/api/v1alpha1"
 
-	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
+	"github.com/giantswarm/logging-operator/pkg/config"
 )
 
 var (
@@ -19,13 +20,16 @@ var (
 	supportAlloyLogs   = semver.MustParse("1.6.0")
 )
 
-const observabilityBundleAppName string = "observability-bundle"
+const (
+	observabilityBundleConfigMapName        = "observability-bundle-logging-extraconfig"
+	observabilityBundleAppName       string = "observability-bundle"
+)
 
 // ObservabilityBundleAppMeta returns metadata for the observability bundle app.
-func ObservabilityBundleAppMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
+func ObservabilityBundleAppMeta(cluster *capi.Cluster) metav1.ObjectMeta {
 	metadata := metav1.ObjectMeta{
-		Name:      lc.AppConfigName(observabilityBundleAppName),
-		Namespace: lc.GetAppsNamespace(),
+		Name:      AppConfigName(cluster, observabilityBundleAppName),
+		Namespace: cluster.GetNamespace(),
 		Labels:    map[string]string{},
 	}
 
@@ -34,10 +38,10 @@ func ObservabilityBundleAppMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
 }
 
 // ObservabilityBundleConfigMapMeta returns metadata for the observability bundle extra values configmap.
-func ObservabilityBundleConfigMapMeta(lc loggedcluster.Interface) metav1.ObjectMeta {
+func ObservabilityBundleConfigMapMeta(cluster *capi.Cluster) metav1.ObjectMeta {
 	metadata := metav1.ObjectMeta{
-		Name:      lc.AppConfigName(lc.GetObservabilityBundleConfigMap()),
-		Namespace: lc.GetAppsNamespace(),
+		Name:      AppConfigName(cluster, observabilityBundleConfigMapName),
+		Namespace: cluster.GetNamespace(),
 		Labels: map[string]string{
 			// This label is used by cluster-operator to find extraconfig. This only works on vintage WCs
 			"app.kubernetes.io/name": observabilityBundleAppName,
@@ -48,9 +52,9 @@ func ObservabilityBundleConfigMapMeta(lc loggedcluster.Interface) metav1.ObjectM
 	return metadata
 }
 
-func GetObservabilityBundleAppVersion(ctx context.Context, client client.Client, lc loggedcluster.Interface) (version semver.Version, err error) {
+func GetObservabilityBundleAppVersion(ctx context.Context, client client.Client, cluster *capi.Cluster) (version semver.Version, err error) {
 	// Get observability bundle app metadata.
-	appMeta := ObservabilityBundleAppMeta(lc)
+	appMeta := ObservabilityBundleAppMeta(cluster)
 	// Retrieve the app.
 	var currentApp appv1.App
 	err = client.Get(ctx, types.NamespacedName{Name: appMeta.GetName(), Namespace: appMeta.GetNamespace()}, &currentApp)
@@ -60,25 +64,29 @@ func GetObservabilityBundleAppVersion(ctx context.Context, client client.Client,
 	return semver.Parse(currentApp.Spec.Version)
 }
 
-func ToggleAgents(ctx context.Context, client client.Client, lc loggedcluster.Interface) error {
+func ToggleAgents(ctx context.Context, client client.Client, cluster *capi.Cluster, cfg config.Config) (*LoggingAgent, error) {
 	logger := log.FromContext(ctx)
 
-	observabilityBundleVersion, err := GetObservabilityBundleAppVersion(ctx, client, lc)
+	observabilityBundleVersion, err := GetObservabilityBundleAppVersion(ctx, client, cluster)
 	if err != nil {
-		return err
+		return nil, err
+	}
+	agent := &LoggingAgent{
+		LoggingAgent:     cfg.DefaultLoggingAgent,
+		KubeEventsLogger: cfg.DefaultKubeEventsLogger,
 	}
 
 	// Enforce promtail as logging agent when observability-bundle version < 1.6.0 because this needs alloy 0.4.0.
-	if observabilityBundleVersion.LT(supportAlloyLogs) && lc.GetLoggingAgent() == LoggingAgentAlloy {
-		logger.Info("Alloy logging agent is not supported by observability bundle, using promtail instead.", "observability-bundle-version", observabilityBundleVersion, "logging-agent", lc.GetLoggingAgent())
-		lc.SetLoggingAgent(LoggingAgentPromtail)
+	if observabilityBundleVersion.LT(supportAlloyLogs) && agent.GetLoggingAgent() == LoggingAgentAlloy {
+		logger.Info("Alloy logging agent is not supported by observability bundle, using promtail instead.", "observability-bundle-version", observabilityBundleVersion, "logging-agent", agent.GetLoggingAgent())
+		agent.SetLoggingAgent(LoggingAgentPromtail)
 	}
 
 	// Enforce grafana-agent as events logger when observability-bundle version < 1.9.0 because this needs alloy 0.7.0.
-	if observabilityBundleVersion.LT(supportAlloyEvents) && lc.GetKubeEventsLogger() == EventsLoggerAlloy {
-		logger.Info("Alloy events logger is not supported by observability bundle, using grafana-agent instead.", "observability-bundle-version", observabilityBundleVersion, "events-logger", lc.GetKubeEventsLogger())
-		lc.SetKubeEventsLogger(EventsLoggerGrafanaAgent)
+	if observabilityBundleVersion.LT(supportAlloyEvents) && agent.GetKubeEventsLogger() == EventsLoggerAlloy {
+		logger.Info("Alloy events logger is not supported by observability bundle, using grafana-agent instead.", "observability-bundle-version", observabilityBundleVersion, "events-logger", agent.GetKubeEventsLogger())
+		agent.SetKubeEventsLogger(EventsLoggerGrafanaAgent)
 	}
 
-	return nil
+	return agent, nil
 }