+++ pkg/resource/agents-toggle/reconciler.go
@@ -6,9 +6,9 @@ import (
 	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 
 	"github.com/giantswarm/logging-operator/pkg/common"
-	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -16,24 +16,24 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
-// Reconciler implements a reconciler.Interface to handle
+// Resource implements a resource.Interface to handle
 // Logging agents toggle: enable or disable logging agents in a given Cluster.
-type Reconciler struct {
+type Resource struct {
 	Client client.Client
 	Scheme *runtime.Scheme
 }
 
 // ReconcileCreate ensure logging agents and events loggers are enabled in the given cluster.
-func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Resource) ReconcileCreate(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("agents toggle create")
 
 	desiredConfigMap := v1.ConfigMap{
-		ObjectMeta: common.ObservabilityBundleConfigMapMeta(lc),
+		ObjectMeta: common.ObservabilityBundleConfigMapMeta(cluster),
 	}
 
 	_, err := controllerutil.CreateOrUpdate(ctx, r.Client, &desiredConfigMap, func() error {
-		config, err := generateObservabilityBundleConfig(ctx, lc)
+		config, err := generateObservabilityBundleConfig(loggingAgent)
 		if err != nil {
 			return errors.WithStack(err)
 		}
@@ -52,12 +52,12 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 }
 
 // ReconcileDelete ensure logging agents and events loggers are disabled for the given cluster.
-func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Resource) ReconcileDelete(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("delete agents toggle config")
 
 	desiredConfigMap := v1.ConfigMap{
-		ObjectMeta: common.ObservabilityBundleConfigMapMeta(lc),
+		ObjectMeta: common.ObservabilityBundleConfigMapMeta(cluster),
 	}
 
 	logger.Info("deleting agents toggle config")