+++ pkg/resource/interface.go
@@ -1,0 +1,20 @@
+package resource
+
+import (
+	"context"
+
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	ctrl "sigs.k8s.io/controller-runtime"
+
+	"github.com/giantswarm/logging-operator/pkg/common"
+)
+
+// Interface provides a resource interface which is the controller core logic
+// for reconciliation loops.
+//
+// An implementation can then be used by a controller to extend its capabilities.
+type Interface interface {
+	ReconcileCreate(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error)
+
+	ReconcileDelete(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error)
+}