+++ pkg/resource/logging-wiring/reconciler.go
@@ -10,31 +10,30 @@ import (
 	apimachineryerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-
-	"github.com/giantswarm/logging-operator/pkg/common"
-	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
-
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/giantswarm/logging-operator/pkg/common"
 )
 
-// Reconciler implements a reconciler.Interface to handle
+// Resource implements a resource.Interface to handle
 // Logging wiring: set or unset the user value configmap created by
 // logging-agents-toggle in the observability bundle.
-type Reconciler struct {
+type Resource struct {
 	Client client.Client
 	Scheme *runtime.Scheme
 }
 
 // ReconcileCreate ensure user value configmap is set in observability bundle
 // for the given cluster.
-func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Resource) ReconcileCreate(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("logging wiring create")
 
 	// Get observability bundle app metadata.
-	appMeta := common.ObservabilityBundleAppMeta(lc)
+	appMeta := common.ObservabilityBundleAppMeta(cluster)
 	// Retrieve the app.
 	logger.Info("logging wiring checking app", "namespace", appMeta.GetNamespace(), "name", appMeta.GetName())
 	var currentApp appv1.App
@@ -48,7 +47,7 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	desiredApp := lc.WireLogging(currentApp)
+	desiredApp := wireLogging(cluster, currentApp)
 	if !reflect.DeepEqual(currentApp, *desiredApp) {
 		logger.Info("logging wiring updating")
 		// Update the app.
@@ -65,12 +64,12 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 
 // ReconcileCreate ensure user value configmap is unset in observability bundle
 // for the given cluster.
-func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Resource) ReconcileDelete(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("logging wiring delete")
 
 	// Get observability bundle app metadata.
-	appMeta := common.ObservabilityBundleAppMeta(lc)
+	appMeta := common.ObservabilityBundleAppMeta(cluster)
 	var currentApp appv1.App
 	err := r.Client.Get(ctx, types.NamespacedName{Name: appMeta.GetName(), Namespace: appMeta.GetNamespace()}, &currentApp)
 	if err != nil {
@@ -82,7 +81,7 @@ func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Inter
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	desiredApp := lc.UnwireLogging(currentApp)
+	desiredApp := unwireLogging(cluster, currentApp)
 	if !reflect.DeepEqual(currentApp, *desiredApp) {
 		logger.Info("logging wiring updating")
 		// Update the app.
@@ -95,4 +94,48 @@ func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Inter
 	logger.Info("logging wiring up to date")
 
 	return ctrl.Result{}, nil
+}
+
+func getWiredExtraConfig(cluster *capi.Cluster) appv1.AppExtraConfig {
+	observabilityBundleConfigMapMeta := common.ObservabilityBundleConfigMapMeta(cluster)
+	return appv1.AppExtraConfig{
+		Kind:      "configMap",
+		Name:      observabilityBundleConfigMapMeta.GetName(),
+		Namespace: observabilityBundleConfigMapMeta.GetNamespace(),
+		Priority:  25,
+	}
+}
+
+// unwireLogging unsets the extraconfig confimap in a copy of the app
+func unwireLogging(cluster *capi.Cluster, currentApp appv1.App) *appv1.App {
+	desiredApp := currentApp.DeepCopy()
+
+	wiredExtraConfig := getWiredExtraConfig(cluster)
+	for index, extraConfig := range currentApp.Spec.ExtraConfigs {
+		if reflect.DeepEqual(extraConfig, wiredExtraConfig) {
+			desiredApp.Spec.ExtraConfigs = append(currentApp.Spec.ExtraConfigs[:index], currentApp.Spec.ExtraConfigs[index+1:]...)
+		}
+	}
+
+	return desiredApp
+}
+
+// wireLogging sets the extraconfig confimap in a copy of the app.
+func wireLogging(cluster *capi.Cluster, currentApp appv1.App) *appv1.App {
+	desiredApp := currentApp.DeepCopy()
+
+	wiredExtraConfig := getWiredExtraConfig(cluster)
+	// We check if the extra config already exists to know if we need to remove it.
+	var containsWiredExtraConfig = false
+	for _, extraConfig := range currentApp.Spec.ExtraConfigs {
+		if reflect.DeepEqual(extraConfig, wiredExtraConfig) {
+			containsWiredExtraConfig = true
+		}
+	}
+
+	if !containsWiredExtraConfig {
+		desiredApp.Spec.ExtraConfigs = append(desiredApp.Spec.ExtraConfigs, wiredExtraConfig)
+	}
+
+	return desiredApp
 }